Rename material-about-form
