+++ src/components/forms/material-about-form/tests/material-about-form.spec.js
@@ -4,12 +4,12 @@ import { useDispatch } from 'react-redux';
 import { MockedProvider } from '@apollo/client/testing';
 import { ThemeProvider } from '@material-ui/styles';
 import { theme } from '../../../app/app-theme/app.theme';
-import MaterialAboutForm from '../material-about-add-form';
+import MaterialAboutForm from '../material-about-form';
 import {
   mockMaterialAboutBlock,
   target,
   files
-} from './material-about-add-form.variables';
+} from './material-about-form.variables';
 import FileReaderMock from '../../../../../__mocks__/fileReaderMock';
 import ImageUploadContainer from '../../../../containers/image-upload-container';
